// src/lib/schemas/detail.ts
export interface NestedCharityClient {
  id: string
  name: string
  logoUrl?: string
}

export interface CharityDetail {
  id: string
  name: string
  description: string
  logoUrl?: string
  contactPhone?: string
  contactEmail?: string
  officialWebsite?: string
  approvalNo?: string
  categories: { id: string; key: CategoryKey; displayName: string }[]
}

export interface DonationDetail {
  id: string
  name: string
  description: string
  logoUrl?: string
  coverImageUrl?: string
  content: string
  raisingApprovalNo?: string
  reliefApprovalNo?: string
  charity: NestedCharityClient
  categories: { id: string; key: CategoryKey; displayName: string }[]
}